refactor: Renamed UrlArrayLoader.ts
