+++ src/colorizer/loaders/UrlArrayLoader.ts
